creating bricks shape by code
BrickManager contains two 2d arrays.
brickGrid - for storing all the bricks in a 2d array
gridPositions - stores the Vector2 positions for each brick according to the 2d grid.

most of the functions are used to dynamically allow the bricks to be plotted in top half of the screen starting from the top left corner with offset adjustments (only in columns) so as to appear centered