add logic to end game once brick touches bottom left prefab which has a rigidbody
currently the way im doing it is by passing the reference of game manager around from the brickgenerator.definegrid function to brickManager to brickServicePool to brickController and calling it inside the brickView when brick triggers collision with bottom left prefab for boundary. Have to look for a better approach without the need for passing game manager reference around. Or probably it can be beneficial in end.